Refactor `CsvConverter` architecture by modularizing type-specific converters
- Remove redundant `convertNumber` method and related tests.
- Extract numeric converters into `NumericConverter` base class for reuse and consistency.
- Move date and boolean converters to separate files.
- Simplify numeric serialization/deserialization logic with centralized decimal formatting.
