+++ src/main/kotlin/io/github/minthem/converter/StringCsvConverter.kt
@@ -1,0 +1,13 @@
+package io.github.minthem.converter
+
+class StringCsvConverter : CsvConverter<String> {
+    override fun deserialize(value: String?): Result<String?> =
+        runCatching {
+            value
+        }
+
+    override fun serialize(value: String?): Result<String?> =
+        runCatching {
+            value
+        }
+}
